pull

+++ scripts/flash_script.sh
@@ -41,23 +41,19 @@ ui_print "************************"
 ui_print "* Magisk v$MAGISK_VER Installer"
 ui_print "************************"
 
-is_mounted /data || mount /data || is_mounted /cache || mount /cache || abort "! Unable to mount partitions"
+is_mounted /data || mount /data || is_mounted /cache || mount /cache
 mount_partitions
-
-find_boot_image
-find_dtbo_image
-
 check_data
 get_flags
+find_boot_image
 
 [ -z $BOOTIMAGE ] && abort "! Unable to detect target image"
 ui_print "- Target image: $BOOTIMAGE"
-[ -z $DTBOIMAGE ] || ui_print "- DTBO image: $DTBOIMAGE"
 
 # Detect version and architecture
 api_level_arch_detect
 
-[ $API -lt 21 ] && abort "! Magisk is only for Lollipop and above (5.0+) (SDK 21+)"
+[ $API -lt 17 ] && abort "! Magisk is only for Android 4.2 and above"
 
 ui_print "- Device platform: $ARCH"
 
@@ -73,18 +69,10 @@ remove_system_su
 
 ui_print "- Constructing environment"
 
-if $DATA; then
-  MAGISKBIN=/data/magisk
-  $DATA_DE && MAGISKBIN=/data/adb/magisk
-  run_migrations
-else
-  MAGISKBIN=/cache/data_bin
-fi
-
 # Copy required files
 rm -rf $MAGISKBIN/* 2>/dev/null
 mkdir -p $MAGISKBIN 2>/dev/null
-cp -af $BINDIR/. $COMMONDIR/. $CHROMEDIR $TMPDIR/bin/busybox $MAGISKBIN
+cp -af $BINDIR/. $COMMONDIR/. $CHROMEDIR $BBDIR/busybox $MAGISKBIN
 chmod -R 755 $MAGISKBIN
 
 # addon.d
@@ -92,9 +80,23 @@ if [ -d /system/addon.d ]; then
   ui_print "- Adding addon.d survival script"
   mount -o rw,remount /system
   ADDOND=/system/addon.d/99-magisk.sh
-  echo '#!/sbin/sh' > $ADDOND
-  echo '# ADDOND_VERSION=2' >> $ADDOND
-  echo 'exec sh /data/adb/magisk/addon.d.sh "$@"' >> $ADDOND
+  cat <<EOF > $ADDOND
+#!/sbin/sh
+# ADDOND_VERSION=2
+
+if [ -f /data/adb/magisk/addon.d.sh ]; then
+  exec sh /data/adb/magisk/addon.d.sh "\$@"
+else
+  OUTFD=\$(ps | grep -v 'grep' | grep -oE 'update(.*)' | cut -d" " -f3)
+  ui_print() { echo -e "ui_print \$1\nui_print" >> /proc/self/fd/\$OUTFD; }
+
+  ui_print "************************"
+  ui_print "* Magisk addon.d failed"
+  ui_print "************************"
+  ui_print "! Cannot find Magisk binaries - was data wiped or not decrypted?"
+  ui_print "! Reflash OTA from decrypted recovery or reflash Magisk"
+fi
+EOF
   chmod 755 $ADDOND
 fi
 
@@ -110,11 +112,21 @@ $BOOTSIGNED && ui_print "- Boot image is signed with AVB 1.0"
 SOURCEDMODE=true
 cd $MAGISKBIN
 
+$IS64BIT && mv -f magiskinit64 magiskinit || rm -f magiskinit64
+
 # Source the boot patcher
 . ./boot_patch.sh "$BOOTIMAGE"
 
 ui_print "- Flashing new boot image"
-flash_image new-boot.img "$BOOTIMAGE" || abort "! Insufficient partition size"
+
+if ! flash_image new-boot.img "$BOOTIMAGE"; then
+  ui_print "- Compressing ramdisk to fit in partition"
+  ./magiskboot cpio ramdisk.cpio compress
+  ./magiskboot repack "$BOOTIMAGE"
+  flash_image new-boot.img "$BOOTIMAGE" || abort "! Insufficient partition size"
+fi
+
+./magiskboot cleanup
 rm -f new-boot.img
 
 if [ -f stock_boot* ]; then
